latest working facial verification

+++ src/main/java/com/attendease/backend/studentModule/dto/request/biometrics/FaceVerificationRequest.java
@@ -1,11 +1,11 @@
 package com.attendease.backend.studentModule.dto.request.biometrics;
 
-import lombok.Data;
-
 import java.util.List;
+import lombok.Data;
 
 @Data
 public class FaceVerificationRequest {
-    private List<Double> uploaded_encoding;
-    private List<Double> reference_encoding;
+
+    private List<Float> uploaded_encoding;
+    private List<Float> reference_encoding;
 }
