+++ src/main/java/com/attendease/backend/studentModule/dto/response/FacialEncodingResponse.java
@@ -1,15 +1,17 @@
 package com.attendease.backend.studentModule.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
 import java.util.List;
+import lombok.Data;
 
 @Data
 public class FacialEncodingResponse {
+
     private boolean success;
     private String message;
+
     @JsonProperty("facialEncoding")
-    private List<Double> facialEncoding;
+    private List<Float> facialEncoding;
+
     private String error;
 }
