+++ src/main/java/com/attendease/backend/studentModule/dto/response/biometrics/FaceEncodingResponse.java
@@ -1,11 +1,11 @@
 package com.attendease.backend.studentModule.dto.response.biometrics;
 
-import lombok.Data;
-
 import java.util.List;
+import lombok.Data;
 
 @Data
 public class FaceEncodingResponse {
+
     private Boolean success;
-    private List<Double> facialEncoding;
+    private List<Float> facialEncoding;
 }
